Added utility for PDF generation
- Bundled PDF generation in utility function.
- Different views can access this utility seperately.
