+++ src/cms/views/pages/page_actions.py
@@ -3,35 +3,29 @@
 Returns:
     [type]: [description]
 """
-import hashlib
 import json
 import logging
 import os
 import uuid
 
 from mptt.exceptions import InvalidMove
-from xhtml2pdf import pisa
 
 from django.contrib import messages
 from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required, permission_required
-from django.contrib.staticfiles import finders
-from django.core.cache import cache
 from django.core.exceptions import PermissionDenied
-from django.db.models import Min
-from django.http import JsonResponse, HttpResponse
+from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404, get_list_or_404
 from django.utils.translation import ugettext as _
 from django.views.static import serve
 from django.http import HttpResponseNotFound
-from django.template.loader import get_template
 
-from backend.settings import WEBAPP_URL, STATIC_URL, MEDIA_URL
-from ...constants import text_directions
+from backend.settings import WEBAPP_URL
 from ...decorators import region_permission_required, staff_required
 from ...forms.pages import PageForm
 from ...models import Page, Language, Region, PageTranslation
 from ...page_xliff_converter import PageXliffHelper, XLIFFS_DIR
+from ...utils.pdf_utils import generate_pdf
 
 logger = logging.getLogger(__name__)
 
@@ -118,58 +112,19 @@ def delete_page(request, page_id, region_slug, language_code):
     )
 
 
-# pylint: disable=unused-argument
-def link_callback(uri, rel):
-    """
-    According to xhtml2pdf documentation (see :doc:`xhtml2pdf:usage`),
-    this function is neccessary for resolving the django static files references.
-    It returns the absolute paths to the files on the file system.
-
-    :param uri: URI that is generated by django template tag 'static'
-    :type uri: str
-
-    :return: The absolute path on the file system according to django's static file settings
-    :rtype: str
-    """
-    if uri.startswith(MEDIA_URL):
-        # Remove the MEDIA_URL from the start of the uri
-        uri = uri[len(MEDIA_URL) :]
-    elif uri.startswith(STATIC_URL):
-        # Remove the STATIC_URL from the start of the uri
-        uri = uri[len(STATIC_URL) :]
-    elif uri.startswith("../"):
-        # Remove ../ from the start of the uri
-        uri = uri[3:]
-    else:
-        logger.warning(
-            "The file %s is not inside the static directories %s and %s.",
-            uri,
-            STATIC_URL,
-            MEDIA_URL,
-        )
-        return uri
-    result = finders.find(uri)
-    if not result:
-        logger.exception(
-            "The file %s was not found in the static directories %s.",
-            uri,
-            finders.searched_locations,
-        )
-    return result
-
-
 @login_required
 @region_permission_required
-# pylint: disable=unused-argument, too-many-locals
+# pylint: disable=unused-argument
 def export_pdf(request, region_slug, language_code):
     """
     Function for handling a pdf export request for pages.
-    The pages were either selected by cms user (see :func:`cms.static.js.pages.page_bulk_action.bulk_action_execute`)
-    or by API request (see :func:`api.v3.pdf_export`)
-    For more information on xhtml2pdf, see :doc:`xhtml2pdf:index`
+    The pages get extracted from request.GET attribute and the request is forwarded to :func:`cms.utils.pdf_utils.generate_pdf`
 
     :param request: Request submitted for rendering pdf document
     :type request: ~django.http.HttpRequest
+
+    :param region_slug: unique region slug
+    :type region_slug: str
 
     :param language_code: bcp47 code of the current language
     :type language_code: str
@@ -180,72 +135,14 @@ def export_pdf(request, region_slug, language_code):
     :rtype: ~django.http.HttpResponse
     """
     region = Region.get_current_region(request)
-    if request.GET.get("pages"):
-        # if the cms users sends an internal request, then retrieve the selected pages
-        page_ids = request.GET.get("pages").split(",")
-        pages = region.pages.filter(archived=False, id__in=page_ids)
-    elif request.GET.get("api"):
-        # if api requests pages
-        pages = request.GET.get("api")
-    pdf_key_list = [region_slug]
-    for page in pages:
-        # retrieve the translation for each page
-        page_translation = page.get_public_translation(language_code)
-        if page_translation:
-            pdf_key_list.append(page_translation.id)
-            pdf_key_list.append(page_translation.last_updated)
-        else:
-            pages = pages.exclude(id=page.id)
-    pdf_key_string = "_".join(map(str, pdf_key_list))
-    pdf_hash_key = hashlib.sha256(bytes(pdf_key_string, "utf-8")).hexdigest()
-    cached_response = cache.get(pdf_hash_key, "has_expired")
-    if cached_response != "has_expired":
-        return cached_response
-
-    amount_pages = pages.count()
-    language = Language.objects.get(code=language_code)
-    template_path = "pages/page_pdf.html"
-    context = {
-        "html2pdf": True,
-        "right_to_left": language.text_direction == text_directions.RIGHT_TO_LEFT,
-        "region": region,
-        "pages": pages,
-        "language": language,
-        "amount_pages": amount_pages,
-    }
-    response = HttpResponse(content_type="application/pdf")
-    if amount_pages == 0:
-        return HttpResponse(_("No valid pages selected for PDF generation."))
-    if amount_pages == 1:
-        # If pdf contains only one page, take its title as filename
-        title = pages.first().get_public_translation(language_code).title
-    else:
-        # If pdf contains multiple pages, check the minimum level
-        min_level = pages.aggregate(Min("level")).get("level__min")
-        # Query all pages with this minimum level
-        min_level_pages = pages.filter(level=min_level)
-        if min_level_pages.count() == 1:
-            # If there's exactly one page with the minimum level, take its title
-            title = min_level_pages.first().get_public_translation(language_code).title
-        else:
-            # In any other case, take the region name
-            title = region.name
-    filename = f"Integreat - {language.translated_name} - {title}.pdf"
-    response["Content-Disposition"] = f'attachment; filename="{filename}"'
-    template = get_template(template_path)
-    html = template.render(context)
-    pisa_status = pisa.CreatePDF(
-        html, dest=response, link_callback=link_callback, encoding="UTF-8"
-    )
-    if pisa_status.err:
-        logger.error(
-            "The following PDF could not be rendered: Region: %s, Language: %s, Pages: %s.",
-            region,
-            language,
-            pages,
-        )
-        return HttpResponse(_("The PDF could not be successfully generated."))
-    cache.set(pdf_hash_key, response, 60 * 60 * 24)
+    # retrieve the selected page ids
+    page_ids = request.GET.get("pages").split(",")
+    # collect the corresponding page objects
+    pages = region.pages.filter(archived=False, id__in=page_ids)
+    # generate PDF document wrapped in a HtmlResponse object
+    response = generate_pdf(region_slug, language_code, region, pages)
+    # offer PDF document for download
+    response["Content-Disposition"] = response["Content-Disposition"] + "; attachment"
     return response
